Move zeek related code to ppl
