+++ ppl/import/PcapFileIcon.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 
-import PcapFileSvg from "../../static/icons/pcap-file.svg"
+import PcapFileSvg from "src/static/icons/pcap-file.svg"
 
 export default function PcapFileIcon(props: any) {
   return <PcapFileSvg className="pcap-file-icon" {...props} />
