+++ ppl/import/ZeekFileIcon.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 
-import ZeekFileSvg from "../../static/icons/zeek-file.svg"
+import ZeekFileSvg from "src/static/icons/zeek-file.svg"
 
 export default function ZeekFileIcon(props: any) {
   return <ZeekFileSvg className="zeek-file-icon" {...props} />
