+++ ppl/zeek/descriptions.ts
