+++ src/js/electron/main.ts
@@ -19,7 +19,7 @@ import {app} from "electron"
 import {handleSquirrelEvent} from "./squirrel"
 import tron from "./tron"
 import path from "path"
-import {ZQD} from "../zqd/zqd"
+import {ZQD} from "ppl/zqd/zqd"
 import electronIsDev from "./isDev"
 import {setupAutoUpdater} from "./autoUpdater"
 import log from "electron-log"
